Refactor formula logic to use decimal; improve sorting UX
Refactored all formula evaluation and storage to use decimal instead of double for improved numeric precision. Updated DataHelper, ExcelService, and FormulaService to support decimal parsing and calculations. Enhanced MainWindow sorting logic with user confirmation and streamlined column selection handling.

+++ FileMappingEngine/Lib/AppManager.cs
@@ -491,7 +491,7 @@ namespace FileMappingEngine.Lib
         {
             foreach (DataRow row in dataTable.Rows)
             {
-                double result = formulaService.Evaluate(formulaTree, row);
+                decimal result = formulaService.Evaluate(formulaTree, row);
 
                 row[targetColumn] = result;
             }
